quick fix for 273 + reset feature
