+++ src/app/actions/course.ts
@@ -89,6 +89,8 @@ export async function coursesBy(
       },
     });
   }
+
+  //console.log("courses: ", courses);
   // Convert to a stored course and sort by number
   const res = courses.map(toStoredCourse);
   res.sort(compareCoursesByNum);
@@ -126,12 +128,14 @@ export async function getSuggestedClasses(
 export async function courseInfo(
   pred: CourseQuery,
 ): Promise<StoredCourse | undefined> {
+  console.log("courseInfo");
   const course = await prisma.course.findFirst({
     where: {
       departmentCode: pred.departmentCode,
       number: pred.number,
     },
   });
+
   if (!course) return undefined;
   return toStoredCourse(course);
 }
